refactor: RaceCount를 객체로 포장
