+++ src/test/java/racingcar/domain/RaceCountTest.java
@@ -8,18 +8,12 @@ import java.util.Arrays;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
-public class RaceTest {
-	private Cars cars;
-
-	@BeforeEach
-	void setUp() {
-		cars = new Cars(Arrays.asList("aa", "bb", "cc"));
-	}
+public class RaceCountTest {
 
 	@ParameterizedTest
 	@ValueSource(strings = {"a", ":", "-", "ㅁ"})
 	void 레이스_횟수_정수(String raceCount) {
-		assertThatThrownBy(() -> new Race(cars, raceCount))
+		assertThatThrownBy(() -> new RaceCount(raceCount))
 				.isInstanceOf(IllegalArgumentException.class)
 				.hasMessageContaining("는 정수가 아닙니다.");
 	}
@@ -27,7 +21,7 @@ public class RaceTest {
 	@ParameterizedTest
 	@ValueSource(strings = {"-1", "-2", "0"})
 	void 레이스_횟수_자연수(String raceCount) {
-		assertThatThrownBy(() -> new Race(cars, raceCount))
+		assertThatThrownBy(() -> new RaceCount(raceCount))
 				.isInstanceOf(IllegalArgumentException.class)
 				.hasMessageContaining("는 자연수가 아닙니다.");
 	}
